Move setup mock script to appcenter module
